+++ src/launchDebugProvider.ts
@@ -60,6 +60,14 @@ export class LaunchDebugProvider implements vscode.DebugConfigurationProvider {
                 void this.launch(this.context, targetUri, userConfig);
             }
         } else if (config && (config.type === 'edge' || config.type === 'msedge')) {
+            void vscode.window.showWarningMessage(
+                `Launch type "${config.type}" is deprecated. Update your launch.json to use "pwa-msedge" instead.`, 'Learn More', 'OK'
+            ).then(value => {
+                if (value === 'Learn More') {
+                    const uri = vscode.Uri.parse('https://code.visualstudio.com/docs/nodejs/browser-debugging');
+                    void vscode.env.openExternal(uri);
+                }
+            });
             const settings = vscode.workspace.getConfiguration(SETTINGS_STORE_NAME);
             if (settings.get('autoAttachViaDebuggerForEdge')) {
                 if (!userConfig.port) {